Transformation composition for part 3
Now, instead of performing two transformations for each of Position and
Orientation Control and summing the delta_thetas together, compose the
Jacobians and delta vectors, and perform a single transformation with
that.

A nice side-effect is that now Pseudo Inverse works when both Position
and Orientation Control is enabled!  Still need to figure out why it
doesn't work when only Orientation Control is enabled...